backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
langchain>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
sentence-transformers>=3.0.0
faiss-cpu>=1.9.0
pypdf>=4.2.0
pydantic>=2.9.0
python-dotenv>=1.0.1
google-generativeai>=0.8.0
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
httpx>=0.27.0
